Improve attainment values in createGraph
